avoid processing epochs multiple times in bitfield queue (#1336)
Given that we're working with pre-quantized epochs, we might end up adding the
same epoch to the `updatedEpochs` list multiple times. This patch fixes this by
iterating twice.

Technically, we could avoid the second map iteration by adding an "is epoch in
quantizedValues" check to the first loop, but doing this in two steps makes this
easier to reason about.